+++ RELEASES.md
@@ -3,6 +3,8 @@
 - Update `dusk-bls12_381` to `0.3.0`.
 - Export `Fr` as `JubJubScalar`
 - Create no-std compatibility via feature.
+- Rename `AffinePoint` to `JubJubAffine`
+- Rename `ExtendedPoint` to `JubJubExtended`
 
 # 0.4.0
 ### Change